backend/pyproject.toml
[project]
name = "study-buddy"
version = "0.1.0"
description = ""
authors = [
    {name = "anekobtw",email = "elmars0608@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic (>=2.12.4,<3.0.0)",
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
